// src/types/profile.ts
export type Profile = {
  username: string;
  name: string;
  bio: string;
  location: string;
  website: string;
  birthDate: string;
  followers: number;
  following: number;
  tweetCount: number;
  joinedDate: string;
  profileImage: string;
  coverImage: string;
  likes: number;
  lists: number;
};